+++ src/app/components/histogram/histogram.component.ts
@@ -6,7 +6,9 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./histogram.component.scss']
 })
 export class HistogramComponent implements OnInit {
-  data = Array.from({length: 1000}, () => Math.floor(Math.random() * 101));
+  selector = 'lib-histogram';
+  data = Array.from({length: 30}, () => Math.floor(Math.random() * 101));
+  dataStr = JSON.stringify(this.data, null, 2);
 
   constructor() { }